changed the ull to size_t and recompiled

+++ divider.c
@@ -7,7 +7,7 @@
  * Return: a number which can divide 'a' with 0 remainder
  */
 
-unsigned long long int divider(unsigned long long int a, unsigned long long int i)
+size_t divider(size_t a, size_t i)
 {
 	if (i * i > a)
 	{
